* Move API into different package. * Refactor.
