+++ tool/automation/framework/app/src/main/kotlin/com/jetbrains/teamcity/Automation.kt
@@ -1,23 +1,23 @@
 /*
  * This Kotlin source file was generated by the Gradle 'init' task.
  */
-package automation
+package com.jetbrains.teamcity
 
 // TODO: Optimize imports
 import kotlinx.serialization.*
 import kotlinx.serialization.json.Json
-
-import DockerImageValidationException
-import automation.common.constants.ValidationConstants
-import automation.docker.hub.models.DockerRegistryAccessor
-import automation.docker.hub.models.DockerRepositoryInfo
-import automation.docker.validation.ImageValidationUtils
-import automation.teamcity.TeamCityUtils
 import kotlinx.cli.ArgParser
 import kotlinx.cli.ArgType
 import kotlinx.cli.Subcommand
 import kotlinx.cli.vararg
 import java.lang.IllegalArgumentException
+import com.jetbrains.teamcity.common.constants.ValidationConstants
+import com.jetbrains.teamcity.docker.DockerImageValidationException
+import com.jetbrains.teamcity.docker.validation.ImageValidationUtils
+import com.jetbrains.teamcity.models.DockerRegistryAccessor
+import com.jetbrains.teamcity.models.DockerRepositoryInfo
+import com.jetbrains.teamcity.teamcity.TeamCityUtils
+
 
 
 /**
@@ -37,10 +37,10 @@ class ValidateImage: Subcommand("validate", "Validate Docker Image") {
 
         // TODO: Move registry URL into config
         val registryAccessor = DockerRegistryAccessor("https://hub.docker.com/v2")
+        val imageNameElements = imageName.split(":")
         // serializing objects
-        // -- use isLateient to sufficiently parse Number arguments (int would be overflown)
-        val registryInfo: DockerRepositoryInfo = Json{ignoreUnknownKeys = true; isLenient = true}.decodeFromString<DockerRepositoryInfo>(registryAccessor.getSize("jetbrains/teamcity-agent", "2022.10-windowsservercore-1809"))
-        TeamCityUtils.reportTeamCityStatistics("SIZE-${ImageValidationUtils.getImageStatisticsId(imageName)}", registryInfo.fullSize)
+        val size = registryAccessor.getSize(imageNameElements[0], imageNameElements[1]);
+        TeamCityUtils.reportTeamCityStatistics("SIZE-${ImageValidationUtils.getImageStatisticsId(imageName)}", size)
         return
 
         // -- report image size to TeamCity
