+++ tool/automation/framework/build.gradle
@@ -17,7 +17,7 @@ repositories {
 sourceSets {
     main {
         kotlin {
-            srcDirs 'app/src/main'
+            srcDirs 'app/src/main/kotlin'
         }
         resources {
             rootProject.files('app/src/main/resources')
@@ -36,5 +36,5 @@ dependencies {
 }
 
 application {
-    mainClass = 'automation.AutomationKt'
+    mainClass = 'com.jetbrains.teamcity.AutomationKt'
 }
